+++ packages/fuels-test-helpers/src/node_types.rs
@@ -1,0 +1,209 @@
+use std::{
+    fmt,
+    io::Write,
+    net::{Ipv4Addr, SocketAddr},
+    path::PathBuf,
+    time::Duration,
+};
+
+pub use fuel_core_chain_config::ChainConfig;
+use fuel_core_chain_config::StateConfig;
+use fuel_core_client::client::FuelClient;
+
+use fuel_types::{BlockHeight, Word};
+use fuels_core::{
+    constants::WORD_SIZE,
+    types::{coin::Coin, message::Message},
+};
+use portpicker::pick_unused_port;
+use serde::{de::Error as SerdeError, Deserializer, Serializer};
+use serde_json::Value;
+use serde_with::{DeserializeAs, SerializeAs};
+use tempfile::NamedTempFile;
+
+use fuels_core::types::errors::Result as FuelResult;
+
+pub const DEFAULT_CACHE_SIZE: usize = 10 * 1024 * 1024;
+
+#[derive(Clone, Debug)]
+pub enum Trigger {
+    Instant,
+    Never,
+    Interval {
+        block_time: Duration,
+    },
+    Hybrid {
+        min_block_time: Duration,
+        max_tx_idle_time: Duration,
+        max_block_time: Duration,
+    },
+}
+
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub enum DbType {
+    InMemory,
+    RocksDb,
+}
+
+#[derive(Clone, Debug)]
+pub struct Config {
+    pub addr: SocketAddr,
+    pub max_database_cache_size: usize,
+    pub database_path: PathBuf,
+    pub database_type: DbType,
+    pub utxo_validation: bool,
+    pub manual_blocks_enabled: bool,
+    pub block_production: Trigger,
+    pub vm_backtrace: bool,
+    pub silent: bool,
+    pub chain_conf: ChainConfig,
+}
+
+impl Config {
+    pub fn local_node() -> Self {
+        Self {
+            addr: SocketAddr::new(Ipv4Addr::new(127, 0, 0, 1).into(), 0),
+            max_database_cache_size: DEFAULT_CACHE_SIZE,
+            database_path: Default::default(),
+            database_type: DbType::InMemory,
+            utxo_validation: true,
+            manual_blocks_enabled: false,
+            block_production: Trigger::Instant,
+            vm_backtrace: false,
+            silent: true,
+            chain_conf: ChainConfig::local_testnet(),
+        }
+    }
+}
+
+impl Default for Config {
+    fn default() -> Self {
+        Self {
+            addr: SocketAddr::new(Ipv4Addr::new(127, 0, 0, 1).into(), 0),
+            max_database_cache_size: DEFAULT_CACHE_SIZE,
+            database_path: Default::default(),
+            database_type: DbType::InMemory,
+            utxo_validation: true,
+            manual_blocks_enabled: false,
+            block_production: Trigger::Instant,
+            vm_backtrace: false,
+            silent: true,
+            chain_conf: ChainConfig::local_testnet(),
+        }
+    }
+}
+
+pub type InternalDaBlockHeight = u64;
+
+pub(crate) struct HexType;
+
+impl<T: AsRef<[u8]>> SerializeAs<T> for HexType {
+    fn serialize_as<S>(value: &T, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        serde_hex::serialize(value, serializer)
+    }
+}
+
+impl<'de, T, E> DeserializeAs<'de, T> for HexType
+where
+    for<'a> T: TryFrom<&'a [u8], Error = E>,
+    E: fmt::Display,
+{
+    fn deserialize_as<D>(deserializer: D) -> Result<T, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        serde_hex::deserialize(deserializer)
+    }
+}
+
+pub mod serde_hex {
+    use std::{convert::TryFrom, fmt};
+
+    use hex::{FromHex, ToHex};
+    use serde::{de::Error, Deserializer, Serializer};
+
+    pub fn serialize<T, S>(target: T, ser: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+        T: ToHex,
+    {
+        let s = format!("0x{}", target.encode_hex::<String>());
+        ser.serialize_str(&s)
+    }
+
+    pub fn deserialize<'de, T, E, D>(des: D) -> Result<T, D::Error>
+    where
+        D: Deserializer<'de>,
+        for<'a> T: TryFrom<&'a [u8], Error = E>,
+        E: fmt::Display,
+    {
+        let raw_string: String = serde::Deserialize::deserialize(des)?;
+        let stripped_prefix = raw_string.trim_start_matches("0x");
+        let bytes: Vec<u8> = FromHex::from_hex(stripped_prefix).map_err(D::Error::custom)?;
+        let result = T::try_from(bytes.as_slice()).map_err(D::Error::custom)?;
+        Ok(result)
+    }
+}
+
+pub(crate) struct HexNumber;
+
+impl SerializeAs<u64> for HexNumber {
+    fn serialize_as<S>(value: &u64, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        let bytes = value.to_be_bytes();
+        serde_hex::serialize(bytes, serializer)
+    }
+}
+
+impl<'de> DeserializeAs<'de, Word> for HexNumber {
+    fn deserialize_as<D>(deserializer: D) -> Result<Word, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let mut bytes: Vec<u8> = serde_hex::deserialize(deserializer)?;
+        match bytes.len() {
+            len if len > WORD_SIZE => {
+                return Err(D::Error::custom(format!(
+                    "value can't exceed {WORD_SIZE} bytes",
+                )));
+            }
+            len if len < WORD_SIZE => {
+                // pad if length < word size
+                bytes = (0..WORD_SIZE - len)
+                    .map(|_| 0u8)
+                    .chain(bytes.into_iter())
+                    .collect();
+            }
+            _ => {}
+        }
+        // We've already verified the bytes.len == WORD_SIZE, force the conversion here.
+        Ok(Word::from_be_bytes(
+            bytes.try_into().expect("byte lengths checked"),
+        ))
+    }
+}
+
+impl SerializeAs<BlockHeight> for HexNumber {
+    fn serialize_as<S>(value: &BlockHeight, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        let number = u32::from(*value) as u64;
+        HexNumber::serialize_as(&number, serializer)
+    }
+}
+
+impl<'de> DeserializeAs<'de, BlockHeight> for HexNumber {
+    fn deserialize_as<D>(deserializer: D) -> Result<BlockHeight, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let number: u64 = HexNumber::deserialize_as(deserializer)?;
+        Ok((number as u32).into())
+    }
+}
